requirements.txt
Flask
SQLAlchemy==1.4.22
pytest==6.2.4
pydantic
python-dateutil
